Arruma tipo do parâmetro da LocadoraDAO.get

+++ src/java/org/dsw/dao/LocadoraDAO.java
@@ -16,7 +16,7 @@ import org.dsw.pojo.Locadora;
 public class LocadoraDAO extends GenericDAO<Locadora>{
 
     @Override
-    public Locadora get(int id) {
+    public Locadora get(Long id) {
         EntityManager em = this.getEntityManager();
         Locadora locadora = em.find(Locadora.class, id);
         em.close();
